feat(owox): added JSDoc to base command
